Can run basic `GET` request on server using pytest
