+++ pytests/__init__.py
@@ -1,0 +1,1 @@
+# This makes a directory a python module
