+++ pytests/test_get_only/test.conf
@@ -1,0 +1,12 @@
+server {
+    listen localhost:9090;
+    root /var/www;
+    server_name webserver-a.org www.webserver-a.org;
+    allow_methods GET;
+    client_max_body_size 4096;
+    error_page 402 /error/402.html;
+    error_page 404 /error/404.html;
+    location / {
+        index html/index.html;
+    }
+}
